requirements.txt
pandas
pyarrow
fastparquet
google-cloud-bigquery
pyspark
jupyterlab

pandas-gbq
